Move copy/paste logic into shared functions
